perf(tests): split icon and theme constants off the data they sit beside
Four extractions, all moving literals verbatim so a non-React consumer can read the
data without the icon COMPONENTS entering its graph:

- currency-theme.constants: colours/gradients/classNames, no @tabler/icons-react.
  currency.constants re-attaches the icons on top.
- file-display-icons: componentTypeConfig, fileFormatConfig, getFileIconConfig --
  the only three exports in file-display-helpers carrying icon components, where
  eight server modules import the rest for pure file-type data.
- donation-goal.constants and training.constants: two constant sets that lived in
  React modules and are read by server schemas.

Note the file-display split is deliberately HARD, with no re-export left in
file-display-helpers. cf-images-utils can re-export from edge-url because the
barrel is the heavy side and the leaf is the light one; here that is reversed, so a
re-export would put @tabler/icons-react straight back into the server graph and
undo the split.

Split out of #3958.
