+++ src/components/Training/Form/TrainingCommon.ts
@@ -23,12 +23,6 @@ import { isDefined } from '~/utils/type-guards';
 export const basePath = '/models/train';
 export const maxSteps = 3;
 
-// nb: these should be proper AIRs now
-export const blockedCustomModels = [
-  'civitai:53761@285757',
-  'urn:air:sd1:checkpoint:civitai:53761@285757',
-];
-
 /**
  * Computes the number of decimal points in a given input using magic math
  */
